refactor(list-manager): split components to seperate files
